Revert "Patches from SolveSpace (#19)"
This reverts commit 8c57f68e76cba83f68e5e136cb2855cd60057328.

+++ src/drw_objects.cpp
@@ -448,13 +448,13 @@ void DRW_LType::parseCode(int code, dxfReader *reader){
 
 //! Update line type
 /*!
-*  Update the size and length of line type according to the path
+*  Update the size and length of line type acording to the path
 *  @author Rallaz
 */
 /*TODO: control max length permited */
 void DRW_LType::update(){
     double d =0;
-    size = (int)path.size();
+    size = path.size();
     for (int i = 0;  i< size; i++){
         d += fabs(path.at(i));
     }
